% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_phosphoproteome.R
\docType{data}
\name{example_phosphoproteome}
\alias{example_phosphoproteome}
\title{Example phosphoproteome.}
\format{A data frame with 6215 rows and 4 variables:
\describe{
  \item{annotation}{Annotation of phosphorylated peptides}
  \item{peptide}{blank - peptides need to be extracted from
  annotation}
  \item{fc}{Fold Change (log2)}
  \item{pval}{P-value for fold-change.}
}}
\usage{
example_phosphoproteome
}
\description{
A dataset containing annotated subtrate sequences derived
from XXX. See original publication for more
details: Engholm-Keller & Waardenberg AJ et al.
}
\keyword{datasets}
