% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphositeplus_human.R
\docType{data}
\name{phosphositeplus_human}
\alias{phosphositeplus_human}
\title{Human kinase-substrates derived from PhosphositePlus.}
\format{A data frame with 11985 rows and 2 variables:
\describe{
  \item{kinase}{human kinase gene symbol}
  \item{substrate}{centered substrate sequence for kinase}
}}
\source{
\url{https://www.phosphosite.org/}
}
\usage{
phosphositeplus_human
}
\description{
A dataset containing human kinases and subtrate sequences. See
original publication for more details: Hornbeck et al.
Nucleic Acids Res. 40:D261-70, 2012
}
\keyword{datasets}
