% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swing.r
\name{swing}
\alias{swing}
\title{Swing statistic}
\usage{
swing(input.data = NULL, pwm.in = NULL, pwm.scores = NULL,
  pseudo.count = 1, p.cut.pwm = 0.05, p.cut.fc = 0.05,
  permutations = 1000, seed = 1234, verbose = FALSE, threads = 1)
}
\arguments{
\item{input.data}{A data.frame of phoshopeptide data. Must contain 4 columns
and the following format must be adhered to. Column 1 - Annotation, Column 2 -
centered peptide sequence, Column 3 - Fold Change [-ve to +ve], Column 4 - p-value [0-1].
This must be the same dataframe used in score.sequences()}

\item{pwm.in}{List of PWMs created using build.pwm()}

\item{pwm.scores}{List of PWM-substrate scores created using
score.seqeuences()}

\item{pseudo.count}{Pseudo-count for avoiding log-zero transformations.
Default: "1"}

\item{p.cut.pwm}{Significance level for determining a significant
kinase-substrate enrichment. Default: "0.05"}

\item{p.cut.fc}{Significance level for determining a significant level of
Fold-change in the phosphoproteomics data. Default: "0.05"}

\item{permutations}{Number of permutations to perform. This will shuffle the kinase-subtrate
edged of the network n times. To not perform permutations and only generate the scores,
set permutations=1 or permutations=FALSE. Default: "1000"}

\item{seed}{This is for reproducible results for permutation testing.
To not use a set seed=NULL. Default: "1234"}

\item{verbose}{Turn verbosity on/off. To turn on, verbose=TRUE. Options are: "TRUE, FALSE".
Default=FALSE}

\item{threads}{Number of processing cores to use. Default: "1"}
}
\value{
A data.table of swing scores.
}
\description{
This function integrates the kinase-substrate predictions,
directionality of phosphopeptide fold change and signficance to assess
local connectivity (swing) of kinase-substrate networks. The final score
is a normalised and weighted score of predicted kinase activity. If permutations
are selected, network node:edges are permutated. P-values will be calculated for
both ends of the distribution of swing scores (positive and negative swing scores).
}
\examples{
## import data
data(example_phosphoproteome)
data(phosphositeplus_human)

## clean up the annotations
## sample 100 data points for demonstration
sample.data <- head(example_phosphoproteome, 100)
annotated.data <- clean.annotation(input.data=sample.data)

## build the PWM models:
set.seed(1)
sample.pwm <- phosphositeplus_human[sample(nrow(phosphositeplus_human), 1000),] 
kinase.pwm.models <- build.pwm(sample.pwm)

## score the PWM - substrate matches
## Using a "random" background, to calculate the p-value of the matches
## Using n=10 for demonstration

pwm.substrate.scores <- score.sequences(input.data = annotated.data, 
                                       pwm.in = kinase.pwm.models,
                                       background = "random", 
                                       n = 10, 
                                       threads = 4)

## Use the pwm.scores and annotated data to predict kinase activity.
## This will permute the network node and edges 10 times for demonstration.
swing.output <- swing(input.data = annotated.data, 
                      pwm.in = kinase.pwm.models, 
                      pwm.scores = pwm.substrate.scores,
                      permutations = 10,
                      threads = 4)

}
